fix(release): share skill platform parsing
